[#1309] Avoid nginx redirecting to URIs with trailing slash
From nginx's documentation at
http://nginx.org/en/docs/http/ngx_http_core_module.html#location:

> If a location is defined by a prefix string that ends with the slash character,
> and requests are processed by one of proxy_pass, fastcgi_pass, uwsgi_pass,
> scgi_pass, or memcached_pass, then the special processing is performed. In
> response to a request with URI equal to this string, but without the trailing
> slash, a permanent redirect with the code 301 will be returned to the requested
> URI with the slash appended.

This means that if we define a location like `location /api/3/ { ... }`, nginx
will automatically redirect requests on `/api/3` to `/api/3/`. As our nginx
instance is accessed via HTTP behind CloudFlare, the redirects will be to the
HTTP version, which breaks AJAX requests and is blocked by some browsers that
want HTTPS-only content.

The correct way would be for nginx to use the content of the
`X-Forwarded-Proto` header, if it exists. Unfortunately, it doesn't, at least
not that I could find. So the fix is to avoid nginx doing these redirects, and
let the application themselves do them as they wish. We can then configure the
applications to handle the `X-Forwarded-Proto` headers.

openspending/openspending#1309